Add mouseon and mouseout handlers.
I'm using a (ToolTip)[http://jqueryui.com/tooltip/] to display information
about events when the user hovers their mouse above a timeline node. When the
mouse moves away from the node, I'd like the tooltip to go away. Currently
this is not possible with timeline.hover(), so I added handlers for the
mouseon and mouseout events.